+++ packages/pages/src/bin/pages.ts
@@ -1,9 +1,9 @@
-import init from "../init/init.js";
-import dev from "../dev/dev.js";
-import preview from "../preview/preview.js";
-import { features } from "../generate/features.js";
-
-const [, , ...args] = process.argv;
+import { initCommandModule } from "../init/init.js";
+import { devCommandModule } from "../dev/dev.js";
+import { generateCommandModule } from "../generate/generate.js";
+import { buildCommandModule } from "../build/build.js";
+import yargs from "yargs";
+import { hideBin } from "yargs/helpers";
 
 // pages requires react and react-dom be installed to function appropriately. If
 // these are not installed in instruct the user to install them.
@@ -18,56 +18,14 @@ const [, , ...args] = process.argv;
   }
 });
 
-const [command] = args;
-
-switch (command) {
-  case "dev":
-    await dev();
-    break;
-  case "preview":
-    preview();
-    break;
-  case "init": {
-    let folderToCreate;
-
-    if (args.length == 2) {
-      folderToCreate = args[1];
-    }
-    await init(folderToCreate || null);
-    break;
-  }
-  case "generate": {
-    const generateTargets = ["features"];
-
-    if (args.length != 2) {
-      process.stdout.write(
-        `Missing a target to generate. Valid values are: ${generateTargets.join(
-          ", "
-        )}\n`
-      );
-      process.exit(1);
-    }
-
-    const target = args[1];
-
-    switch (target) {
-      case "features": {
-        await features();
-        break;
-      }
-      default: {
-        process.stdout.write(
-          `Target for 'generate' command is invalid: ${target}. Valid values are: ${generateTargets.join(
-            ", "
-          )}\n`
-        );
-        process.exit(1);
-      }
-    }
-    break;
-  }
-  default: {
-    process.stdout.write(`Command not found: ${command}\n`);
-    process.exit(1);
-  }
-}
+yargs(hideBin(process.argv))
+  .scriptName("pages")
+  .command(devCommandModule)
+  .command(initCommandModule)
+  .command(generateCommandModule)
+  .command(buildCommandModule)
+  .demandCommand()
+  .version(false)
+  .strict()
+  .help()
+  .parse();