+++ packages/pages/src/components/analytics/provider.tsx
@@ -22,12 +22,17 @@ export function AnalyticsProvider(
     enableTrackingCookie,
     enableDebugging,
     templateData,
+    pageDomain,
   } = props;
 
   const analyticsRef = useRef<AnalyticsMethods | null>(null);
 
   if (analyticsRef.current === null) {
-    analyticsRef.current = new Analytics(templateData, requireOptIn);
+    analyticsRef.current = new Analytics(
+      templateData,
+      requireOptIn,
+      pageDomain
+    );
   }
 
   const analytics = analyticsRef.current;
@@ -36,8 +41,11 @@ export function AnalyticsProvider(
     analytics.enableTrackingCookie();
   }
 
-  const enableDebuggingDefault =
-    debuggingParamDetected() || process?.env?.NODE_ENV === "development";
+  let enableDebuggingDefault = debuggingParamDetected();
+  if (typeof process !== "undefined") {
+    enableDebuggingDefault =
+      enableDebuggingDefault || process.env?.NODE_ENV === "development";
+  }
   analytics.setDebugEnabled(enableDebugging ?? enableDebuggingDefault);
 
   return (