+++ packages/pages/src/dev/client/entry-client.tsx
@@ -9,28 +9,29 @@ const hydrate = async () => {
   };
 
   // Can't use string interpolation here so src/templates is hardcoded
-  const templates = import.meta.glob("/src/templates/*.(jsx|tsx)");
+  const templates = import.meta.glob("/src/templates/**/*.(jsx|tsx)");
+
+  const getTemplateName = (path: string) =>
+    path.split("/").pop()?.split(".")[0] || "index";
 
   const routes: Route[] = Object.keys(templates).map((path) => {
     return {
       // get the filename from the path and remove its extension, default to index
-      name: path.split("/").pop()?.split(".")[0] || "index",
+      name: getTemplateName(path),
       path,
       getComponent: templates[path],
     };
   });
 
   /**
-   * Get the templateFilename from the template. See {@link ./ssr/serverRenderRoute.ts}.
+   * Get the templatePath from the template. See {@link ./ssr/serverRenderRoute.ts}.
    */
-  const templateFilename = (window as any)._RSS_TEMPLATE_;
-  const templateFilenameWithoutSuffix = templateFilename?.split(".")[0];
-
-  const template = routes.find(
-    (route) => route.name === templateFilenameWithoutSuffix
+  const templatePath = (window as any)._RSS_TEMPLATE_PATH_;
+  const template = routes.find((route) =>
+    templatePath.endsWith(route.path)
   ) || {
-    name: templateFilename?.split(".")[0],
-    path: `/src/templates/${templateFilename}`,
+    name: getTemplateName(templatePath),
+    path: templatePath,
     getComponent: function (): Promise<any> {
       throw new Error("Function not implemented.");
     },