+++ packages/pages/src/dev/server/middleware/serverRenderRoute.ts
@@ -11,6 +11,8 @@ import {
 import { ProjectStructure } from "../../../common/src/project/structure.js";
 import { TemplateModuleInternal } from "../../../common/src/template/internal/types.js";
 import templateBase from "../public/templateBase.js";
+import { lookup } from "mime-types";
+import { getTemplateFilepaths } from "../../../common/src/template/internal/getTemplateFilepaths.js";
 
 type Props = {
   vite: ViteDevServer;
@@ -26,9 +28,18 @@ export const serverRenderRoute =
 
       const { feature, entityId, locale } = urlToFeature(url);
 
+      const templateFilepaths = getTemplateFilepaths(
+        projectStructure.scopedTemplatesPath
+          ? [
+              projectStructure.scopedTemplatesPath,
+              projectStructure.templatesRoot,
+            ]
+          : [projectStructure.templatesRoot]
+      );
       const templateModuleInternal = await featureNameToTemplateModuleInternal(
         vite,
-        feature
+        feature,
+        templateFilepaths
       );
       if (!templateModuleInternal) {
         console.error(
@@ -76,7 +87,7 @@ export const serverRenderRoute =
         `<head>
             <script type="text/javascript">
               window._RSS_PROPS_ = ${JSON.stringify(props)};
-              window._RSS_TEMPLATE_ = '${templateModuleInternal.filename}';
+              window._RSS_TEMPLATE_PATH_ = '${templateModuleInternal.path}';
               window._RSS_LANG_ = '${getLang(headConfig, props)}';
             </script>
             ${headConfig ? renderHeadConfigToString(headConfig) : ""}
@@ -105,6 +116,6 @@ const getContentType = (
   // TODO: once custom headers are supported at the template level use that instead,
   // with a fallback to the current logic.
   const path = templateModuleInternal.getPath(props);
-  const ext = path.includes(".") ? path.split(".").pop() : "";
-  return ext || "text/html";
+
+  return lookup(path) || "text/html";
 };