+++ packages/pages/src/dev/server/ssr/constants.ts
@@ -1,6 +1,7 @@
-export const TEMPLATE_PATH = "src/templates";
 export const STREAM_DATA_CHUNK_BEGIN = "{";
 export const STREAM_DATA_CHUNK_END = "}";
+export const STREAM_DATA_CHUNK_BEGIN_MULTIPLE = "[";
+export const STREAM_DATA_CHUNK_END_MULTIPLE = "]";
 export const CLI_BOILERPLATE_BETA_MESSAGE =
   "This is a beta version of the Yext Command Line Interface";
 export const UPGRADE_MESSAGE_LINE_BEGIN = "Please upgrade the Yext CLI";