+++ packages/pages/src/init/init.ts
@@ -1,7 +1,12 @@
 import fs from "fs";
 import { runGenerate } from "./generate/interface.js";
+import { CommandModule } from "yargs";
 
-export default async (folderToCreate: string | null) => {
+interface InitArgs {
+  folderToCreate?: string;
+}
+
+const handler = async ({ folderToCreate }: InitArgs) => {
   if (folderToCreate) {
     await fs.promises.mkdir(folderToCreate);
     process.chdir(folderToCreate);
@@ -16,4 +21,18 @@ export default async (folderToCreate: string | null) => {
     }
     runGenerate();
   }
+};
+
+export const initCommandModule: CommandModule<unknown, InitArgs> = {
+  command: "init",
+  describe:
+    "Clones a pages starter repo, runs `npm install`, and performs the first build.",
+  builder: (yargs) => {
+    return yargs.option("folderToCreate", {
+      describe: "Destination folder path to create the repo",
+      type: "string",
+      demandOption: false,
+    });
+  },
+  handler,
 };