+++ packages/pages/CHANGELOG.md
@@ -1,3 +1,42 @@
+#### 1.0.0-beta.9 (2022-10-04)
+
+##### Bug Fixes
+
+* **components:**  remove `process?` optional chaining (#237) (b48875ba)
+
+#### 1.0.0-beta.8 (2022-10-04)
+
+##### Performance Improvements
+
+* **plugin:**  dedupe the css file list to avoid unnecessary imports (#235) (8d96efc9)
+
+#### 1.0.0-beta.7 (2022-09-30)
+
+##### Chores
+
+* **components:**
+  *  bump yext/analytics version (#231) (373ac02b)
+  *  upgrade storybook to use vite (#217) (16fd089c)
+*  remove vitepress and use storybook (#221) (5d757d06)
+*  add prepare script for @yext/pages (#214) (0e2d0f67)
+
+##### Documentation Changes
+
+* **components:**  add address stories (#227) (82fa43f0)
+
+##### New Features
+
+* **components:**
+  *  add Map component (#171) (c41d3239)
+  *  enable debugging for Analytics in dev mode (#226) (02327faf)
+* **dev:**  allow render-only templates to work (#219) (ac1eab96)
+
+##### Bug Fixes
+
+* **components:**  set sizes attribute for images (#232) (3c843b3e)
+*  features command when using top-level await (#230) (9f32e197)
+* **dev:**  remove final slash from static page urls on index (#220) (6aac7bc9)
+
 #### 1.0.0-beta.6 (2022-09-08)
 
 ##### Chores