+++ packages/plugins/CHANGELOG.md
@@ -1,3 +1,19 @@
+#### 1.0.0-beta.3 (2022-10-04)
+
+##### Bug Fixes
+
+* **pages-plugins:**  handle json object input (#234) (abb00a68)
+
+#### 1.0.0-beta.2 (2022-09-30)
+
+##### Chores
+
+*  update READMEs for Pages Plugins (#216) (43bc00e6)
+
+##### Bug Fixes
+
+* **plugin:**  fix slug connector pagination (#233) (e6e1da97)
+
 #### 1.0.0-beta.1 (2022-09-12)
 
 ##### New Features